+++ Fabric/src/main/java/fuzs/puzzleslib/api/event/v1/FabricEvents.java
@@ -3,6 +3,7 @@ package fuzs.puzzleslib.api.event.v1;
 import fuzs.puzzleslib.api.event.v1.core.FabricEventFactory;
 import net.fabricmc.fabric.api.event.Event;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.AnvilMenu;
 
 /**
  * Events originally found on Forge in the <code>net.minecraftforge.event</code> package.
@@ -16,4 +17,8 @@ public final class FabricEvents {
      * Fires at the end of {@link Player#tick()}.
      */
     public static final Event<PlayerTickEvents.End> PLAYER_TICK_END = FabricEventFactory.create(PlayerTickEvents.End.class);
+    /**
+     * Called before a result item is generated from the two input slots in an anvil in {@link AnvilMenu#createResult()}.
+     */
+    public static final Event<AnvilUpdateCallback> ANVIL_UPDATE = FabricEventFactory.createResult(AnvilUpdateCallback.class);
 }