+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v5.0.9-1.19.3] - 2023-03-24
+### Added
+- Added a few new events
+- Added some more helper methods to `AbstractModelProvider`
+### Fixed
+- Fixed `CreativeModeTab$Builder` instances not being built on Fabric
+
 ## [v5.0.8-1.19.3] - 2023-03-23
 ### Added
 - Added a few new events